refactor(SceneByVariableRepeaterGrid): Pull out "groupBy" logic out of the class (#109)
